Create EntryRespectingComboBox
When the user types something directly into a combo box that has an editable text field, when the combo box loses focus, if the typed text is semantically similar to an existing entry, then the box tries to overwrite the user's typed value with the existing value.  This combo box class attempts to prevent that.