(WIP) Create ProfileParamsValidator.cs
Created the ProfileParamsValidator class and made it implement the System.Web.UI.IValidator interface.  Am going to be breaking off from this project for a quick second to actually define the same interface in a new 'xyLOGIX.Validators' library that will then be added to this solution -- for semantic consistency.