Create module to manage the paths to multiple "root folders"
A "root folder" is the starting point of the directory tree in which find/replace operations begin.  We want to build out a feature such that the user has the option to set a path to ANY folder in the Root Folder Path drop down in the GUI, not just one know to contain "the current solution."  Then, the software can scan the directory tree of whichever folder the user chooses, doing the same find/replace operation for ALL projects in ALL solutions in that folder's hierarchy.