(WIP) Update MainWindowPresenter.cs
As part of an ongoing effort to allow the use of this tool to operate on directory trees that do not contain .sln files, we had placed a call to the ProgramFlowHelper.StartDebugger method here.  However, the requirement is to move this debugger break call to another layer of the application as we continue testing.