Update MFR.GUI.csproj
Make Post-Build event deploy to either C:\Program Files or C:\Program Files (x86), depending on the bitness of the project.

This is important, because of the fact that this tool needs to potentially store long lists of files in memory, the 64-bit version can address more RAM than can the 32-bit version.