Update VisualStudioSolutionService.cs
Add the following properties:
1. FolderToSearch

Added a bit of extra logging in the ContainsLoadedSolutions()  method to ensure that the logic is performing as desired.

Made the ContainsLoadedSolutions method fall back on the value of the FolderToSearch property in case its 'folder' parameter is blank, or the 'folder' parameter does not contain the fully-qualified pathname of a folder that exists on the disk (for robustness' sake and semantic fluency).

Added the ContainLoadedSolutions method.  For call fluency.  It simply is an alias of the ContainsLoadedSolutions method.

Updated the code formatting and performed a CodeMaid/ReSharper code cleanup on the file.