Create the 'MFR.Params.LoadedSolutions' module
Provides POCOs that package a IVisualStudioSolution interface together with stats such as whether it is inside or out of the root directory (i.e., the directory that the Project File Renamer strats its operations in) directory tree, etc.

This module is a component of the xyLOGIX Project File Renamer software system.

---
Copyright © 2023 by xyLOGIX, LLC.  All rights reserved.