Update OptionsDialog.Designer.cs
* Created a 'Push Changes To Remote When Done' check box on the Git tab

Turning off this check box will turn off the capability to push the changes to the remote when a project file renamer operation is complete.  This is alsready done when the MFR app is run in headless mode (such as during the creation of a software module by the Multi-Project Template Wizard).  The ability of git.exe to push changes quickly to a remote has degraded severely in recent months, making pushes a time-consuming operation.  Best to have the option to not automatically push when a project-file-rename operation is complete.