Update IProjectFileRenamerConfig.cs
* Added the ErrantProcessesList property

This property is a list of strings, each of which is the .exe filename of a process that is to be killed prior to the execution of the Rename Files in Folder operation and the Rename Sub Folder(s) operation.

This list is currently hardcoded in the body of the KillErrantProcesses method; however, we are making it user-configurable according to best practices.