Create the 'MFR.CommandLine.Handlers' module
Provides objects and functionality to handle the various use-cases when invalid input is sent over the command line; either a root directory is blank or missing, does not contain .sln file(s) in its directory tree, or there are multiple instance(s) of Visual Studio open but none of them have the Solution(s) found in the target directory tree, loaded.  We want to develop various user-assistance strategy(ies) to help the user with these use-cases.

This module is a component of the xyLOGIX Project File Renamer software system.

---
Copyright © 2021-25 by xyLOGIX, LLC  All rights reserved.